+++ proadmin-android/app/src/main/java/com/sasd13/proadmin/android/util/adapter/itf2bean/user/UserPreferenceAdapterI2B.java
@@ -10,7 +10,7 @@ public class UserPreferenceAdapterI2B implements IAdapter<UserPreferenceBean, Us
     public UserPreference adapt(UserPreferenceBean s) {
         UserPreference t = new UserPreference();
 
-        t.setId(Integer.valueOf(s.getId()));
+        t.setId(Long.valueOf(s.getId()));
         t.setValue(s.getValue());
         t.setCategory(s.getCategory());
         t.setName(s.getName());
